Refactor NamespacedId to public API
