+++ runtime/engine/src/main/java/io/aklivity/zilla/runtime/engine/namespace/NamespacedId.java
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations
  * under the License.
  */
-package io.aklivity.zilla.runtime.engine.internal.stream;
+package io.aklivity.zilla.runtime.engine.namespace;
 
 public final class NamespacedId
 {
